chore(*): refactored tests, refactored errors, added docs, added loggingFn definition to httpClient protocol
